// src/lib/enums/user.enums.ts
export enum AccessLevel {
    OWNER = 'owner',
    ADMIN = 'admin',
    MANAGER = 'manager',
    SUPERVISOR = 'supervisor',
    USER = 'user',
    DEVELOPER = 'developer',
    SUPPORT = 'support',
}

export enum Department {
    SUPPORT = 'support',
    ENGINEERING = 'engineering',
    SALES = 'sales',
    MARKETING = 'marketing',
    FINANCE = 'finance',
    HR = 'hr',
    LEGAL = 'legal',
    OPERATIONS = 'operations',
} 